Added an extra mapping function for osx canadian keyboard
Because emacs sort of overrides some part of the OSX keyboard.  By that I mean
that the '/' key does 'é' when the CAN keyboard is selected.  In the CAN
keyboard, you can do ALT+é to get a '/' but emacs doesn't see "insert '/'" it
sees 'M-é' and says that this is undefined.