Add github.com/google/vim-searchindex plugin as submodule
Shows '<n>/<tot> <expr>' in the bottom left where
<n> = index of the result from the start of the file
<tot> = total number of search results
<expr> = the searched expression

The simple line 'set shortmess-=S' removes the 'S' from the 'shortmess'
setting causing Vim to display [n/tot] in the bottom right however, this
shows [n/>99] if there are more than 99 results.  This 99 is hardcoded
and cannot be changed.  Hence using a plugin.